[Refactor/Caffe2] Refactor caffe2 tensor-filter subplugin to remove .c/.cc wrapper
This commit refactors caffe2 tensor-filter subplugin to remove .c/.cc
wrapper.
