+++ ext/nnstreamer/tensor_filter/meson.build
@@ -130,10 +130,7 @@ if get_option('enable-pytorch')
 endif
 
 if get_option('enable-caffe2')
-  filter_sub_caffe2_sources = [
-    'tensor_filter_caffe2.c',
-    'tensor_filter_caffe2_core.cc'
-  ]
+  filter_sub_caffe2_sources = ['tensor_filter_caffe2.cc']
 
   nnstreamer_filter_caffe2_sources = []
   foreach s : filter_sub_caffe2_sources
